requirement.txt
nltk
speechrecognition 
googletrans 
pyttsx3 
openai 
python-dotenv
